More flexible category manager

+++ core/classes/models/Category.php
@@ -6,6 +6,9 @@ use core\classes\Model;
 
 class Category extends Model {
 
+	protected $link_type = ''; // either 'foreign-key' or 'link-table'
+	protected $link_from = ''; // table this is linked from via foreign key
+
 	protected $children = [];
 
 	public function addChild(Category $category) {
@@ -31,7 +34,12 @@ class Category extends Model {
 
 	public function delete() {
 		// remove all link records
-		$sql = "DELETE FROM ".$this->table."_link WHERE ".$this->primary_key."=".$this->id;
+		if ($this->link_type == 'link-table') {
+			$sql = "DELETE FROM ".$this->table."_link WHERE ".$this->primary_key."=".$this->id;
+		}
+		elseif ($this->link_type == 'foreign-key') {
+			$sql = "UPDATE ".$this->link_from." SET ".$this->table."_id=NULL WHERE ".$this->primary_key."=".$this->id;
+		}
 		$this->database->executeQuery($sql);
 
 		parent::delete();
